pyproject.toml
[project]
name = "ai-as-me"
version = "0.1.0"
description = "自进化AI数字分身系统 - Agent CLI编排与Soul注入"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "requests>=2.28.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
ai-as-me = "ai_as_me.cli_main:cli"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
